+++ R/univariate_1freq_plot.R
@@ -27,7 +27,12 @@ ggplot2::autoplot
 #' autoplot(x, sort = TRUE, dec.mark = ".", color_bg = "mediumseagreen")
 #'
 #' @export
-autoplot.univ_freq <- function(x, background = TRUE, labels = TRUE, sort = FALSE, sort_manual = NULL, dec.mark = ",", color = "dodgerblue", color_bg = "#E7B800", label_width = 10){
+autoplot.univ_freq <- function(x, background = TRUE, labels = TRUE, sort = FALSE, sort_manual = NULL, dec.mark = ",",
+                               color = "dodgerblue", color_bg = "#E7B800", label_width = 10){
+
+  if (!requireNamespace("ggplot2", quietly = TRUE)) {
+    stop("ggplot2 is needed for this function to work. Install it via install.packages(\"ggplot2\")", call. = FALSE)
+  }
 
   df <- x$df
   xvar <- x$xvar